% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\alias{clean.contact_survey}
\title{Clean contact survey data}
\usage{
\method{clean}{contact_survey}(x, participant.age.column = "part_age", ...)
}
\arguments{
\item{x}{a "contact_survey" object, created with \code{\link[=as_contact_survey]{as_contact_survey()}}.}

\item{participant.age.column}{the column in \code{x$participants} containing participants' age}

\item{...}{ignored}
}
\value{
a cleaned survey in the correct format
}
\description{
Cleans survey data to work with the 'contact_matrix' function
}
\examples{
\dontrun{
#' # not run because it requires an internet connection
current_surveys <- list_surveys()
polymod_url <- subset(current_surveys, grepl("POLYMOD", title))[["url"]]
# "https://doi.org/10.5281/zenodo.3874557"
polymod <- get_survey(polymod_url)
polymod
cleaned <- clean(polymod) # not really necessary as the 'polymod' data set has already been cleaned
}
}
