% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsurveys-dir.R
\name{contactsurveys_dir}
\alias{contactsurveys_dir}
\title{Default directory for persistent storage for contact surveys}
\usage{
contactsurveys_dir()
}
\value{
the active \code{contactsurveys} directory.
}
\description{
Use \code{contactsurveys_dir()} to view the default storage location for files
that are downloaded. The \code{\link[=download_survey]{download_survey()}} function downloads files into
\code{contactsurveys_dir()}. This uses a directory that is specific to your
operating system, powered by the base function, \code{\link[tools:userdir]{tools::R_user_dir()}}. You
can override this by setting an environment variable, \code{CONTACTSURVEYS_HOME}.
The functions \code{\link[=list_surveys]{list_surveys()}} and \code{\link[=download_survey]{download_survey()}} will use this default
directory. You can also specify the \code{directory} argument in these functions
in place of the default value, \code{contactsurveys_dir()}. This approach has been
borrowed from Carl Boettiger's \code{neonstore} R package.
}
\examples{

contactsurveys_dir()

## Override with an environment variable:
Sys.setenv(CONTACTSURVEYS_HOME = tempdir())
contactsurveys_dir()
## Unset
Sys.unsetenv("CONTACTSURVEYS_HOME")

}
