% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-contactsurvey-files.R
\name{delete_survey}
\alias{delete_survey}
\title{Delete files for a given survey}
\usage{
delete_survey(survey)
}
\arguments{
\item{survey}{Survey, as used in \code{\link[=download_survey]{download_survey()}}.}
}
\value{
nothing, deleted files
}
\description{
For when you want to delete files associate with a given survey.
}
\examples{
\dontrun{
peru_doi <- "https://doi.org/10.5281/zenodo.1095664"
delete_survey(peru_doi)
}
}
\seealso{
\code{\link[=delete_contactsurveys_dir]{delete_contactsurveys_dir()}} \code{\link[=delete_contactsurveys_files]{delete_contactsurveys_files()}}
}
