% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.R
\name{.get_survey}
\alias{.get_survey}
\title{Internal function to get survey data}
\usage{
.get_survey(survey, ...)
}
\arguments{
\item{survey}{a DOI or url to get the survey from, or a for a "contact_survey" object, created
with \code{\link[=as_contact_survey]{as_contact_survey()}}, object (in which case only cleaning is done).}

\item{...}{extra options (currently not used)}
}
\description{
Internal function to get survey data
}
\keyword{internal}
