% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{get_citation}
\alias{get_citation}
\title{Citation for a survey}
\usage{
get_citation(x)
}
\arguments{
\item{x}{a character vector of surveys to cite}
}
\value{
citation as bibentry
}
\description{
Gets a full citation for a "contact_survey" object, created
with \code{\link[=as_contact_survey]{as_contact_survey()}}.
}
\examples{
\dontrun{
# not run because it requires an internet connection
current_surveys <- list_surveys()
polymod_url <- subset(current_surveys, grepl("POLYMOD", title))[["url"]]
# "https://doi.org/10.5281/zenodo.3874557"
polymod <- get_survey(polymod_url)
polymod
citation <- get_citation(polymod)
print(citation)
print(citation, style = "bibtex")
}
}
