% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_online}
\alias{is_online}
\title{Are we online?}
\usage{
is_online(host = "captive.apple.com")
}
\arguments{
\item{host}{host URL, default is "captive.apple.com".}
}
\value{
logical - TRUE if online, FALSE otherwise
}
\description{
Are we online?
}
\examples{
if (is_online()){
  is_online()
}
}
