% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-contactsurvey-files.R
\name{ls_contactsurveys}
\alias{ls_contactsurveys}
\title{List files in contactsurveys cache}
\usage{
ls_contactsurveys(dir = contactsurveys_dir())
}
\arguments{
\item{dir}{Directory to list. Default is \code{\link[=contactsurveys_dir]{contactsurveys_dir()}}.}
}
\value{
Character vector of file paths under \code{dir} (absolute paths).
}
\description{
It can be handy to see which files are available, sometimes to return to
a clean slate and delete them via the \verb{delete_contactsurveys_*()} functions.
}
\examples{
ls_contactsurveys()
}
\seealso{
\code{\link[=delete_contactsurveys_files]{delete_contactsurveys_files()}} \code{\link[=delete_survey]{delete_survey()}}
\code{\link[=delete_contactsurveys_dir]{delete_contactsurveys_dir()}} \code{\link[=download_survey]{download_survey()}} \code{\link[=contactsurveys_dir]{contactsurveys_dir()}}
}
