% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-citation.R
\name{print.csbib}
\alias{print.csbib}
\title{Print a csbib citation}
\usage{
\method{print}{csbib}(x, ...)
}
\arguments{
\item{x}{A \code{"csbib"} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a csbib citation
}
