% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_survey.R
\name{store_reference}
\alias{store_reference}
\title{Extracts meta-data and repository info from a zen4R::ZenodoRecord object}
\usage{
store_reference(records, survey_dir)
}
\arguments{
\item{records}{ZenodoRecord object; the object to parse information from}

\item{survey_dir}{file path; location to store the meta-data file}
}
\value{
file path; file path to the JSON file with meta-data and link(s)
}
\description{
Extracts meta-data and repository info from a zen4R::ZenodoRecord object
}
\note{
internal
}
