% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{survey_countries}
\alias{survey_countries}
\title{List all countries contained in a survey}
\usage{
survey_countries(survey, country.column = "country", ...)
}
\arguments{
\item{survey}{a DOI or url to get the survey from, or a for a "contact_survey" object, created
with \code{\link[=as_contact_survey]{as_contact_survey()}}, object (in which case only cleaning is done).}

\item{country.column}{column in the survey indicating the country}

\item{...}{further arguments for \code{\link[=get_survey]{get_survey()}}}
}
\value{
list of countries
}
\description{
List all countries contained in a survey
}
\examples{
\dontrun{
# not run because it requires an internet connection
current_surveys <- list_surveys()
polymod_url <- subset(current_surveys, grepl("POLYMOD", title))[["url"]]
# "https://doi.org/10.5281/zenodo.3874557"
polymod <- get_survey(polymod_url)
polymod
survey_countries(polymod)
}
}
