% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsurveys-package.R
\docType{package}
\name{contactsurveys-package}
\alias{contactsurveys}
\alias{contactsurveys-package}
\title{contactsurveys: Download Contact Surveys for Use in Infectious Disease Modelling}
\description{
Download, cache, and manage social contact survey data from the social contact data community on Zenodo (\url{https://zenodo.org/communities/social_contact_data}) for use in infectious disease modelling. Provides functions to list available surveys, download survey files with automatic caching, and retrieve citations. Contact survey data describe who contacts whom in a population and are used to parameterise age-structured transmission models, for example via the 'socialmixr' package. The surveys available include those from the POLYMOD study (Mossong et al. (2008) \doi{10.1371/journal.pmed.0050074}) and other social contact data shared on Zenodo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiforecasts/contactsurveys}
  \item \url{http://epiforecasts.io/contactsurveys/}
  \item Report bugs at \url{https://github.com/epiforecasts/contactsurveys/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})
}

Other contributors:
\itemize{
  \item Hugo Gruson (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
}

}
\keyword{internal}
