% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsurveys-package.R
\docType{package}
\name{contactsurveys-package}
\alias{contactsurveys}
\alias{contactsurveys-package}
\title{contactsurveys: Download Contact Surveys for Use in Infectious Disease Modelling}
\description{
Provides methods for downloading contact surveys that are then used in infectious disease modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiforecasts/contactsurveys}
  \item \url{http://epiforecasts.io/contactsurveys/}
  \item Report bugs at \url{https://github.com/epiforecasts/contactsurveys/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})
}

Other contributors:
\itemize{
  \item Hugo Gruson (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
}

}
\keyword{internal}
