% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsurveys-package.R
\docType{package}
\name{contactsurveys-package}
\alias{contactsurveys}
\alias{contactsurveys-package}
\title{contactsurveys: Download Contact Surveys for use in Infectious Disease Modelling}
\description{
Provides methods for downloading contact surveys that are then used in infectious disease modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiforecasts/contactsurveys}
  \item \url{http://epiforecasts.io/contactsurveys/}
  \item Report bugs at \url{https://github.com/epiforecasts/contactsurveys/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}

Authors:
\itemize{
  \item Lander Willem
  \item Hugo Gruson
}

Other contributors:
\itemize{
  \item Maria Bekker-Nielsen Dunbar [contributor]
  \item Carl A. B. Pearson [contributor]
  \item Sam Clifford \email{sj.clifford@gmail.com} [contributor]
  \item Christopher Jarvis [contributor]
  \item Alexis Robert \email{alexis.robert@lshtm.ac.uk} [contributor]
  \item Niel Hens [contributor]
  \item Pietro Coletti [collector, data manager]
}

}
\keyword{internal}
