% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-contactsurvey-files.R
\name{delete_contactsurveys_files}
\alias{delete_contactsurveys_files}
\alias{delete-files}
\alias{delete_contactsurveys_dir}
\title{Delete files in contactsurveys directory}
\usage{
delete_contactsurveys_files(dir = contactsurveys_dir())

delete_contactsurveys_dir(dir = contactsurveys_dir())
}
\arguments{
\item{dir}{directory to list files to delete from}
}
\value{
nothing, deletes files.
}
\description{
Delete files in contactsurveys directory
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
delete_contactsurveys_files()
\dontshow{\}) # examplesIf}
}
