% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_survey.R
\name{download_survey}
\alias{download_survey}
\title{Download a survey from its Zenodo repository}
\usage{
download_survey(
  survey,
  directory = contactsurveys_dir(),
  verbose = TRUE,
  overwrite = FALSE,
  timeout = 60
)
}
\arguments{
\item{survey}{A DOI of a survey, (see \code{\link[=list_surveys]{list_surveys()}}). If a HTML link is
given, the DOI will be isolated and used.}

\item{directory}{Directory of where to save survey files. The default value
is to use the directory for your system, based on \code{\link[tools:userdir]{tools::R_user_dir()}}.
You can specify your own directory, or set an environment variable,
\code{CONTACTSURVEYS_HOME}, see \code{\link[=Sys.setenv]{Sys.setenv()}} or \link{Renviron}.}

\item{verbose}{Whether downloads should be echoed to output. Default TRUE.}

\item{overwrite}{If files should be overwritten if they already exist.
Default TRUE.}

\item{timeout}{A numeric value specifying timeout in seconds. Default 60
seconds.}
}
\value{
a vector of filenames, where the surveys were downloaded
}
\description{
Downloads survey data
}
\examples{
\donttest{
list_surveys()
peru_survey <- download_survey("https://doi.org/10.5281/zenodo.1095664")
}
}
\seealso{
\code{\link[=list_surveys]{list_surveys()}}
}
