% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_survey.R
\name{download_survey}
\alias{download_survey}
\title{Download a survey from its Zenodo repository}
\usage{
download_survey(
  survey,
  directory = tempdir(),
  verbose = TRUE,
  overwrite = FALSE,
  timeout = 60
)
}
\arguments{
\item{survey}{A URL (see \code{\link[=list_surveys]{list_surveys()}})}

\item{verbose}{Whether downloads should be echoed to output. Default TRUE.}

\item{overwrite}{If files should be overwritten if they already exist.
Default TRUE.}

\item{timeout}{A numeric value specifying timeout in seconds. Default 60
seconds.}

\item{dir}{a directory to save the files to; if not given, will save to a
temporary directory}
}
\value{
a vector of filenames that can be used with \code{\link[=load_survey]{load_survey()}}
}
\description{
Downloads survey data
}
\examples{
\dontrun{
list_surveys()
peru_survey <- download_survey("https://doi.org/10.5281/zenodo.1095664")
}
}
\seealso{
\code{\link[=load_survey]{load_survey()}}
}
