% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{get_citation}
\alias{get_citation}
\title{Citation for a survey}
\usage{
get_citation(doi, style = "apa", ...)
}
\arguments{
\item{doi}{A DOI.}

\item{style}{Optional. Citation style.}

\item{...}{Passed to \code{zen4R::get_citation()}.}
}
\value{
A citation. Prints a message with citation instructions.
}
\description{
Gets a full citation from a DOI
}
\examples{
\dontrun{
# not run because it requires an internet connection
polymod_url <- "https://doi.org/10.5281/zenodo.3874557"
citation <- get_citation(polymod)
}
}
