% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_surveys}
\alias{list_surveys}
\title{List all surveys available for download}
\usage{
list_surveys(directory = contactsurveys_dir(), overwrite = FALSE)
}
\arguments{
\item{directory}{Directory of where to save survey files. The default value
is to use the directory for your system using \code{\link[=contactsurveys_dir]{contactsurveys_dir()}}, which
uses \code{\link[tools:userdir]{tools::R_user_dir()}}. This effectively caches the data. You can
specify your own directory, or set an environment variable,
\code{CONTACTSURVEYS_HOME}, see \code{\link[=Sys.setenv]{Sys.setenv()}} or \link{Renviron} for more detail.
If this argument is set to something other than \code{\link[=contactsurveys_dir]{contactsurveys_dir()}}, a
warning is issued to discourage unnecessary downloads outside the persistent
cache.}

\item{overwrite}{If files should be overwritten if they already exist.
Default FALSE}
}
\value{
character vector of surveys
}
\description{
List all surveys available for download
}
\examples{
\dontrun{
list_surveys()
}
}
